Debug

+++ src/ssl_remote_sensing/pretext_tasks/simclr/augmentation.py
@@ -39,4 +39,4 @@ class Augment:
         )
 
     def __call__(self, x):
-        return self.train_transform(x), self.train_transform(x)
+        return self.train_transform(x).float(), self.train_transform(x).float()
